control_node/cbdash/pyproject.toml
[tool.poetry]
name = "cbdash"
version = "0.1.0"
description = ""
authors = ["Andrea Zanotti <andreazanottifo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
flask-cors = "^3.0.9"
PyYAML = "^5.3.1"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
python-dotenv = "^0.15.0"
pytest-flask = "^1.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
